Backend_2/requirements.txt
# Core Flask & Web Framework
Flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2

# Database & ORM
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
Alembic==1.12.1
psycopg2-binary==2.9.7

# Authentication & Security
Flask-Login==0.6.3
Flask-Session==0.4.0
bcrypt==4.0.1
cryptography==41.0.7
PyJWT==2.8.0

# API & Data Handling
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
requests==2.31.0

# AI & Machine Learning
google-generativeai==0.3.2
pandas==2.1.3
numpy>=1.26.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Data Analysis & Visualization
scipy==1.11.3
statsmodels==0.14.0
plotly==5.17.0

# Utilities & Development
python-dotenv==1.0.0
gunicorn==21.2.0
blinker==1.6.3

# Export & Reporting
openpyxl==3.1.2
XlsxWriter==3.1.9
reportlab==4.0.6
